Refactor to object methods - may be overreaching
Note: ?? output must be a band (or instrument player list?) that can be passed to a callback
Each control (performer list) needs to have current-instrument context (i.e. what the instrument is and who’s playing it), somewhere to store that back to, and a callback to post-process that data.
Lists need to be aware of: who’s playing this instrument, who’s playing other instruments, who’s already filled their pending allocation.

“Band” is contextual to either self-submission page or admin manage/edit form - which is why those were held in outer scope - not possible with two different scopes.